jpa작업중 - 신형주

+++ src/main/java/com/green/campingsmore/order/payment/model/InsPayInfoDto.java
@@ -3,6 +3,8 @@ package com.green.campingsmore.order.payment.model;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonProperty.Access;
 import com.green.campingsmore.entity.PayType;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -10,15 +12,14 @@ import lombok.NoArgsConstructor;
 import java.util.List;
 
 @Data
-
 public class InsPayInfoDto {
     @JsonProperty(access = JsonProperty.Access.READ_ONLY)
     private Long iuser;
     private Long ireserve;
     private String address;
     private String addressDetail;
-    private Long totalPrice;
-    private Long shippingPrice;
+    private Integer totalPrice;
+    private Integer shippingPrice;
     private String shippingMemo;
     private PayType type;
     private List<PayDetailInfoVo> purchaseList;
